services/web/requirements.txt
bcrypt==4.0.1
Flask==2.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
Werkzeug==2.2.3
gunicorn
flask_jwt_extended